feat(createPedido):Crea el endpoint de crear pedido

+++ marly-handmade-backend/src/main/java/com/marly/handmade/domain/producto/data/ProductoRequest.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 public record ProductoRequest(
         @NotNull(message = "El nombre no puede estar vacio")
         String nombre,
-        Float precio,
+        Double precio,
         Integer stock,
         String fotoPrincipal,
         String fotoSecundario,
